# ai-specs/specs/api-spec.yml
openapi: "3.1.0"
info:
  title: SDD Application API
  version: "1.0.0"
  description: REST API for the SDD application

servers:
  - url: http://localhost:8000
    description: Local development

paths:
  /api/v1/companies:
    get:
      tags: [companies]
      summary: List companies (paginated)
      operationId: listCompanies
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated list of companies
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/CompanyResponse"
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
    post:
      tags: [companies]
      summary: Create a new company
      operationId: createCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyCreate"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "409":
          description: Company name already exists
        "422":
          description: Validation error
    delete:
      tags: [companies]
      summary: Bulk delete companies
      operationId: bulkDeleteCompanies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteRequest"
      responses:
        "204":
          description: Companies deleted
        "422":
          description: Validation error (empty IDs list)

  /api/v1/companies/{company_id}:
    parameters:
      - name: company_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [companies]
      summary: Get a single company by ID
      operationId: getCompany
      responses:
        "200":
          description: Company found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "404":
          description: Company not found
    put:
      tags: [companies]
      summary: Update a company
      operationId: updateCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "404":
          description: Company not found
        "409":
          description: Company name already exists
        "422":
          description: Validation error
    delete:
      tags: [companies]
      summary: Delete a single company
      operationId: deleteCompany
      responses:
        "204":
          description: Company deleted
        "404":
          description: Company not found

components:
  schemas:
    CompanyCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
        street:
          type: string
          maxLength: 255
          nullable: true
        city:
          type: string
          maxLength: 100
          nullable: true
        state:
          type: string
          maxLength: 100
          nullable: true
        zip_code:
          type: string
          maxLength: 20
          nullable: true
        country:
          type: string
          maxLength: 100
          nullable: true
        revenue:
          type: number
          default: 0
        expenses:
          type: number
          default: 0
        employees:
          type: integer
          default: 0
        clients:
          type: integer
          default: 0

    CompanyUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
        street:
          type: string
          maxLength: 255
          nullable: true
        city:
          type: string
          maxLength: 100
          nullable: true
        state:
          type: string
          maxLength: 100
          nullable: true
        zip_code:
          type: string
          maxLength: 20
          nullable: true
        country:
          type: string
          maxLength: 100
          nullable: true
        revenue:
          type: number
          default: 0
        expenses:
          type: number
          default: 0
        employees:
          type: integer
          default: 0
        clients:
          type: integer
          default: 0

    CompanyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        street:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip_code:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        revenue:
          type: number
        expenses:
          type: number
        profit:
          type: number
          description: Computed as revenue - expenses
        employees:
          type: integer
        clients:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    BulkDeleteRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object